Added comment in code regarding firewalld and local interfaces
I can't actually seem to find a way to replace the "source address" property with something that would whitelist a specific local interface. If someone is more familiar with firewalld, feel free to PR changes to the command below so we can whitelist a full local interface to do connects.